separate metatool logic from exiftool logic
